requirements.txt
pandas
requests
beautifulsoup4
python-docx
openpyxl
streamlit